thinsnap: add snapshot field to LV CR
Add optional `snapshot` field to the LogicalVolume
Custom Resource.
This field will only be set in case of a snapshot
logical volume request.
It contains information about the snapshot like it's
type(thick/thin);datasource and accessType; set  `ro` in case of
VolumeSnapshots and `rw` for PVC Restore.(restored PVC is also a
thin-snapshot internally!).

Signed-off-by: Yug Gupta <yuggupta27@gmail.com>